Fixed all screens, imports, and dependencies.

+++ app/src/main/java/com/elfennani/aniwatch/data/local/dao/SessionDao.kt
@@ -9,7 +9,7 @@ import com.elfennani.aniwatch.data.local.models.LocalSession
 @Dao
 interface SessionDao {
     @Insert
-    suspend fun insertSession(session: LocalSession):Int
+    suspend fun insertSession(session: LocalSession):Long
 
     @Query("DELETE FROM LocalSession WHERE id=:id")
     suspend fun deleteSessionById(id:Int)
